+++ react-app/src/components/common/index.js
@@ -18,3 +18,5 @@ export { default as DatabaseActions } from './DatabaseActions';
 export { default as DownloadProgressModal } from './DownloadProgressModal';
 export { default as StorageQuotaModal } from './StorageQuotaModal';
 export { default as ServiceWorkerStatus } from './ServiceWorkerStatus';
+export { default as BrowserSupportStatus } from './BrowserSupportStatus';
+export { default as OfflineCompatibilityBanner } from './OfflineCompatibilityBanner';